+++ packages/falcon-shop-extension/src/types.ts
@@ -453,12 +453,7 @@ export type CheckoutDetailsInput = {
   data?: object;
 };
 
-export type PlaceOrderResult = PlaceOrderSuccessfulResult | PlaceOrder3dSecureResult;
-
-export type PlaceOrderSuccessfulResult = {
-  orderId: string;
-  orderRealId: string;
-};
+export type PlaceOrderResult = Order | PlaceOrder3dSecureResult;
 
 export type PlaceOrder3dSecureResult = {
   url: string;